Fixed failing test for searching where (#4744)

+++ ivy/functional/backends/tensorflow/searching.py
@@ -51,7 +51,7 @@ def where(
     out: Optional[Union[tf.Tensor, tf.Variable]] = None,
 ) -> Union[tf.Tensor, tf.Variable]:
     x1, x2 = ivy.promote_types_of_inputs(x1, x2)
-    return tf.experimental.numpy.where(condition, x1, x2)
+    return tf.cast(tf.experimental.numpy.where(condition, x1, x2), x1.dtype)
 
 
 # Extra #
